Backend/E-Taas/core/config.py
from dotenv import load_dotenv
import os
load_dotenv()

class Settings:
    DATABASE_URL: str = os.getenv("DATABASE_URL")
    SECRET_KEY: str = os.getenv("SECRET_KEY")
    ALGORITHM: str = os.getenv("ALGORITHM")
    JWT_EXPIRATION_MINUTES: int = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES"))
    PAYMONGO_SECRET_KEY: str = os.getenv("PAYMONGO_SECRET_KEY")
    FIREBASE_CREDENTIALS_PATH: str = os.getenv("FIREBASE_CREDENTIALS_PATH")

settings = Settings()

