Backend/requirements.txt
fastapi[all]
uvicorn[standard]
pydantic
sqlalchemy[asyncio]
databases
alembic
asyncpg
pytest
pytest-asyncio
httpx
python-dotenv
python-multipart
bcrypt==3.2.2
PyJWT
email-validator
passlib[bcrypt]
cloudinary
locust
slowapi
firebase-admin